openstack-status: report full service name
While the abbreviated name might have been cleaner,
it was a little ambiguous and precluded direct
copying of the service names for other commands.
Also the use of a global $svc variable introduced
some awkward coupling which wasn't properly updated
when this command was updated to support swift and cinder.