openstack-db: make --drop more encompassing
Given that we check for users @ any host,
we should drop all users no matter what host.
It was seen for example that in certain cases that
the Host for  acreated user would default to '127.0.0.1'
which doesn't match with 'localhost' in mysql.

Querying the users to drop, also allows use to remove
the GRANT hack that was being used to ensure users existed,
before the drop was run.